move posts to blog-posts component
