+++ app/templates/posts.hbs
@@ -1,28 +1,1 @@
-<div class="container-fluid">
-	<div class="row">
-	  <!-- blog post listing -->
-		<div class="col-xs-5">
-		  <button {{action 'createPost'}}>Create</button> {{input type="text" value=query placeholder="Filter Posts" autofocus="autofocus"}}
-			<table class='table'>
-				<thead>
-					<tr><th>Recent Posts</th></tr>
-				</thead>
-				<tbody>
-				{{#each pagedContent as |post|}}
-				<tr><td>
-					{{#link-to 'post' post}}
-						{{post.title}} <small class='muted'>by {{post.author}}</small>
-					{{/link-to}}
-				</td></tr>
-				{{/each}}
-				</tbody>
-			</table>
-			{{page-numbers content=pagedContent}}
-		</div>
-
-		<!-- blog post content -->
-		<div class="col-xs-7">
-			{{outlet}}
-		</div>
-	</div>
-</div>
+{{#blog-posts posts=model page=page perPage=perPage query=query createAction="createPost"}}{{outlet}}{{/blog-posts}}
