+++ tests/integration/components/blog-posts-test.js
@@ -1,0 +1,27 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+
+moduleForComponent('blog-posts', 'Integration | Component | blog posts', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  assert.expect(2);
+
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{blog-posts}}`);
+
+  assert.equal(this.$().text(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#blog-posts}}
+      template block text
+    {{/blog-posts}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});
